Api/requirements.txt
astroid==1.6.3
attrs==17.4.0
click==6.7
coverage==4.5.1
Flask==0.12.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
nose==1.3.7
pluggy==0.6.0
py==1.5.3
PyJWT==1.6.1
pylint==1.8.4
pytest==3.5.1
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
